Expanding tests to cover scalings

+++ tests/test_line.py
@@ -116,7 +116,7 @@ class TestLineCollectionOperations:
         )
         assert np.allclose(
             scaled_lines.continuum.value,
-            simple_line_collection.continuum.value
+            simple_line_collection.continuum.value.astype(np.float64)
             * 2.0
             / simple_line_collection.nu.value,
         ), (
@@ -198,7 +198,7 @@ class TestLineCollectionOperations:
         )
         assert np.allclose(
             scaled_lines.continuum.value,
-            lines.continuum.value * 2.0 / lines.nu.value,
+            lines.continuum.value.astype(np.float64) * 2.0 / lines.nu.value,
         ), (
             "Scaled continuum doesn't match "
             f"{scaled_lines.continuum.value} !="
